refactor frontend to have components subdirectories for organization/readability/clean code
