+++ frontend/src/__tests__/authentication/InvalidEmail.test.js
@@ -2,7 +2,7 @@ import React from 'react'
 import { render, screen, fireEvent } from '@testing-library/react'
 import { ThemeProvider, createTheme } from '@mui/material/styles'
 import { MemoryRouter, useNavigate } from 'react-router-dom'
-import InvalidEmail from '../../components/Auth/InvalidEmail'
+import InvalidEmail from '../../components/authentication/InvalidEmail'
 
 // Need to wrap the component in this because it uses navigate from react-router-dom
 const renderWithTheme = (ui) => {
