+++ frontend/src/__tests__/profiles/FacultyProfileView.test.js
@@ -1,9 +1,9 @@
 import React from 'react'
 import { render, screen, waitFor, fireEvent } from '@testing-library/react'
-import FacultyProfileView from '../components/FacultyProfileView'
+import FacultyProfileView from '../../components/profiles/FacultyProfileView'
 import { MemoryRouter, useNavigate } from 'react-router-dom'
 import { ThemeProvider, createTheme } from '@mui/material/styles'
-import { getFacultyCurrent_expected } from '../resources/mockData'
+import { getFacultyCurrent_expected } from '../../resources/mockData'
 
 // Need to wrap the component in this because it uses navigate from react-router-dom
 const renderWithTheme = (ui) => {
