+++ frontend/src/components/MainLayout.js
@@ -7,13 +7,13 @@
 import React, { useState, useEffect, useCallback } from 'react'
 import { Box, CircularProgress } from '@mui/material'
 import MainAccordion from './MainAccordion'
-import PostDialog from './PostDialog'
-import TitleButton from './TitleButton'
-import SearchBar from './SearchBar'
-import ViewProfile from './ViewProfile'
-import Sidebar from './Sidebar'
+import PostDialog from './posts/PostDialog'
+import TitleButton from './navigation/TitleButton'
+import SearchBar from './filtering/SearchBar'
+import ViewProfile from './profiles/ViewProfile'
+import Sidebar from './filtering/Sidebar'
 import PropTypes from 'prop-types'
-import ViewButton from './ViewButton'
+import ViewButton from './filtering/ViewButton'
 import { fetchStudentsUrl, fetchProjectsUrl, viewStudentsFlag, viewProjectsFlag } from '../resources/constants'
 // import { mockStudents, mockResearchOps } from '../resources/mockData'
