+++ frontend/src/components/MajorAccordion.js
@@ -7,7 +7,7 @@
 import React from 'react'
 import { Accordion, AccordionSummary, AccordionDetails, Typography, Box } from '@mui/material'
 import ExpandMoreIcon from '@mui/icons-material/ExpandMore'
-import PostList from './PostList'
+import PostList from './posts/PostList'
 import PropTypes from 'prop-types'
 import { viewProjectsFlag } from '../resources/constants'
