+++ frontend/src/components/authentication/Logout.js
