+++ frontend/src/components/authentication/RequireAuth.js
