+++ frontend/src/components/authentication/RequireProfile.js
