+++ frontend/src/components/authentication/RequireStudentProfile.js
