+++ frontend/src/components/filtering/SearchBar.js
