+++ frontend/src/components/filtering/Sidebar.js
